requirements.txt
pandas==1.3.5
protobuf==4.21.1
six==1.16.0
tqdm==4.64.0
google-cloud-translate==2.0.1
google-cloud-storage==2.4.0